initial styling enhancements + routing imporvements as per pr feedback
